make runner script generic

+++ run_keras_example.sh
@@ -1,10 +1,15 @@
 #!/bin/bash
+if [ $# -lt 1 ]; then
+    echo usage $(basename $0) "[ script name ]"
+    exit 1
+fi
+
+
 source utils.sh
 source constants.py
 
 log "starting up"
 
-log_dir=/git/tensorboard/addition_rnn_graph
 log "log_dir = "${log_dir}
 
 killall -HUP tensorboard
@@ -12,4 +17,4 @@ rm -rf ${log_dir}
 mkdir -p ${log_dir}
     
 tensorboard --logdir=${log_dir} --port=8888 &
-python addition_rnn.py 2>&1 | tee addition_rnn.log
+python $1 2>&1 | tee addition_rnn.log
